Refactor: move repository classes into extra package
