+++ src/main/java/techcourse/myblog/repository/ArticleRepository.java
@@ -1,6 +1,7 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import techcourse.myblog.domain.Article;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
