+++ src/test/java/techcourse/myblog/domain/ArticleRepositoryTest.java
@@ -3,6 +3,7 @@ package techcourse.myblog.domain;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import techcourse.myblog.repository.ArticleRepository;
 
 import java.util.Optional;
